# packages/frontend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "德州扑克游戏API",
    "version": "1.0.0",
    "description": "德州扑克游戏后端API文档"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "avatar": {
            "type": "string",
            "description": "头像"
          },
          "chips": {
            "type": "number",
            "description": "筹码数量"
          },
          "real_name_verified": {
            "type": "boolean",
            "description": "实名认证状态"
          },
          "verification_status": {
            "type": "string",
            "enum": [
              "pending",
              "verified",
              "rejected"
            ],
            "description": "认证状态"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "username",
          "chips",
          "real_name_verified",
          "verification_status",
          "created_at",
          "updated_at"
        ]
      },
      "RegisterUserInput": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginUserInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          }
        },
        "required": [
          "password"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/users/register": {
      "post": {
        "summary": "用户注册",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "用户注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "注册失败"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "用户登录",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserInput"
              },
              "example": {
                "username": "exampleuser",
                "password": "password123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "登录失败"
          }
        }
      }
    },
    "/users/profile": {
      "get": {
        "summary": "获取用户资料",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取用户资料成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权或获取用户资料失败"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      },
      "put": {
        "summary": "更新用户资料",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "avatar": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新用户资料成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "更新用户资料失败"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/users/stats": {
      "get": {
        "summary": "获取用户统计数据",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取用户统计数据成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "stats": {
                      "type": "object",
                      "properties": {
                        "gamesPlayed": {
                          "type": "number"
                        },
                        "gamesWon": {
                          "type": "number"
                        },
                        "winRate": {
                          "type": "number"
                        },
                        "totalChips": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "获取用户统计数据失败"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    }
  },
...
}